requirements.txt
# Core ML
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
shap>=0.43
joblib

# API
fastapi>=0.100
uvicorn[standard]>=0.23
pydantic>=2.0
httpx

# Dashboard
streamlit>=1.28
plotly>=5.18

# GenAI
google-genai>=1.0

# Data
pyarrow

# Dev
python-dotenv
pytest
